Fix for constant settigns rewrite
I added a NeedsSave property to base settings class which defaults to
true; the load process will set it to false and reset it to true if a
property is not loaded. This in turn will force a resave of the file if
it exists. I also added code to checkout the file before save following
the model used by the normal file create routines of the VSManager code.