Async Action Support
This adds support for Async actions on an async controller. This feature
can be toggled on/off via a SupportAsyncActions settings, which is false
by default. Enabling the setting will create additional methods in the
partial for an async controller that provides a basic ActionResult with
the properties of the Action.